+++ src/index.ts
@@ -1,3 +1,5 @@
+import { int } from '@kobayami/number-types';
+
 /**
  * Returns `true` if the given value is not `undefined`.
  * Returns `false` otherwise.
@@ -20,6 +22,21 @@ export function assertPresent<T>(value: T | undefined, defaultValue?: T): T {
     if (isPresent(value)) return value;
     if (isPresent(defaultValue)) return defaultValue;
     throw noSuchElement();
+}
+
+/**
+ * Returns a value and erases `undefined` from its type, if the value is not `undefined`.
+ * Otherwise, executes `onElse` and returns its return value.
+ *
+ * This function is similar to `assertPresent`, but `onElse` is executed only when `value === undefined`,
+ * while `defaultValue` is always evaluated, which may not be wanted for complex expressions.
+ *
+ * @param value value to check
+ * @param onElse else branch to execute if `value === undefined`
+ */
+export function presentOrElse<T, U>(value: T | undefined, onElse: () => U): T | U {
+    if (isPresent(value)) return value;
+    return onElse();
 }
 
 /**
@@ -48,6 +65,21 @@ export function assertValue<T>(value: T | null | undefined, defaultValue?: T): T
 }
 
 /**
+ * Returns a value and erases `null` and `undefined` from its type, if the value is not `null` or `undefined`.
+ * Otherwise, executes `onElse` and returns its return value.
+ *
+ * This function is similar to `assertValue`, but `onElse` is executed only when `value` is `null` or `undefined`,
+ * while `defaultValue` is always evaluated, which may not be wanted for complex expressions.
+ *
+ * @param value value to check
+ * @param onElse else branch to execute if `value === null` or `value === undefined`
+ */
+export function valueOrElse<T, U>(value: T | null | undefined, onElse: () => U): T | U {
+    if (isValue(value)) return value;
+    return onElse();
+}
+
+/**
  * Returns `true` if the given arrays are of the same size.
  * Returns `false` otherwise.
  *
@@ -55,7 +87,6 @@ export function assertValue<T>(value: T | null | undefined, defaultValue?: T): T
  * @param b second array
  */
 export function isEqualLength(a: any[], b: any[]): boolean {
-    for (let i = 0; i < 10; i++);
     return a.length === b.length;
 }
 
@@ -66,7 +97,7 @@ export function isEqualLength(a: any[], b: any[]): boolean {
  * @param a first array
  * @param b second array
  */
-export function assertEqualLength<T>(a: T[], b: T[]): number {
+export function assertEqualLength<T>(a: T[], b: T[]): int {
     if (isEqualLength(a, b)) return a.length;
     throw illegalArgument('Arrays must have same size');
 }
@@ -78,7 +109,7 @@ export function assertEqualLength<T>(a: T[], b: T[]): number {
  * @param array the array
  * @param index index to check
  */
-export function isValidIndex(array: any[], index: number): boolean {
+export function isValidIndex(array: any[], index: int): boolean {
     return index >= 0 && index < array.length;
 }
 
@@ -89,7 +120,7 @@ export function isValidIndex(array: any[], index: number): boolean {
  * @param array the array
  * @param index index to check
  */
-export function assertValidIndex(array: any[], index: number): number {
+export function assertValidIndex(array: any[], index: int): int {
     if (isValidIndex(array, index)) return index;
     throw illegalArgument('Index out of range');
 }
@@ -98,13 +129,13 @@ export function assertValidIndex(array: any[], index: number): number {
  * Returns `true` if a range of indices entirely resides within the given array.
  * Returns `false` otherwise.
  *
- * Please note that the empty tange `start = end = array.length` is explicitly allowed.
+ * Please note that the empty range `start = end = array.length` is explicitly allowed.
  *
  * @param array the array
  * @param start start of the range (inclusive)
  * @param end end of the range (exclusive)
  */
-export function isValidRange(array: any[], start: number, end: number): boolean {
+export function isValidRange(array: any[], start: int, end: int): boolean {
     return start >= 0 && end >= start && end <= array.length;
 }
 
@@ -112,14 +143,14 @@ export function isValidRange(array: any[], start: number, end: number): boolean 
  * Asserts that if a range of indices entirely resides within the given array.
  * Throws an error otherwise.
  *
- * Please note that the empty tange `start = end = array.length` is explicitly allowed.
+ * Please note that the empty range `start = end = array.length` is explicitly allowed.
  *
  * @param array the array
  * @param start first index in the range
  * @param end first index after the range
  */
 
-export function assertValidRange(array: any[], start: number, end: number): void {
+export function assertValidRange(array: any[], start: int, end: int): void {
     if (isValidRange(array, start, end)) return;
     throw illegalArgument('Invalid range');
 }
@@ -172,4 +203,15 @@ export function invalidState(message?: string): Error {
  */
 export function unsupportedOperation(message?: string): Error {
     return new Error(assertValue(message, 'Unsupported operation'));
+}
+
+/**
+ * Returns an `Error` to indicate that
+ * an operation is supported by the API, but not yet implemented.
+ * This usually means work in progress.
+ *
+ * @param message error message
+ */
+export function notImplemented(message?: string): Error {
+    return new Error(assertValue(message, 'Not implemented'));
 }