+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## 1.1.0
+
+- New guards: `presentOrElse`, `valueOrElse`
+- New exception: `notImplemented`
+- Switched index types from `number` to `int` (see [`kobayami/number-types`](https://kobayami.github.io/number-types/docs/modules.html))
+
 ## 1.0.1
 
 - Fixed typo in `README.ME` usage example